--- my-app-frontend/src/ArchivedItem.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-const ArchivedItem = () => {
-    return ( 
-        <div></div>
-     );
-}
- 
-export default ArchivedItem;
-
